+++ source/daq/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_firmware_component(
     NAME daq
     LINKER_SCRIPT "STM32F407VGTx_FLASH"
-    LIBS "CMSIS_F407;PHAL_F407;FREERTOS_F407;HEARTBEAT"
+    LIBS "CMSIS_F407;PHAL_F407;FREERTOS_F407;SDIO;HEARTBEAT"
 )
